Change type of MB_CUR_MAX and MB_CUR_MAX_L() to size_t.
POSIX requires that MB_CUR_MAX expands to an expression of type size_t.
It currently expands to an int. As these are already macros, don't
change the underlying type of these functions. There is no ned to touch
those.

Differential Revision:	https://reviews.freebsd.org/D6645


Former-commit-id: 495961efc834bdda8599614b6b05a240cf5bf279

+++ include/stdlib.h
@@ -210,7 +210,7 @@ begin_define
 define|#
 directive|define
 name|MB_CUR_MAX
-value|(___mb_cur_max())
+value|((size_t)___mb_cur_max())
 end_define
 
 begin_function_decl
